Move Value Object Classes
Move value object classes into their own namespace "Version".
