+++ src/IP.php
@@ -2,8 +2,10 @@
 
 namespace Darsyn\IP;
 
+use Darsyn\IP\Version\Multi;
+
 trigger_error(
-    '"Darsyn\IP\IP" is deprecated and will be removed in the next major version; use "Darsyn\IP\Multi" instead.',
+    '"Darsyn\IP\IP" is deprecated and will be removed in the next major version; use "Darsyn\IP\Version\Multi" instead.',
     E_USER_DEPRECATED
 );
